fix(battleground): Fix BG premature ending and bot teleport crash
- Fix BGs ending prematurely due to "not enough players" by adding
  MonitorActiveBattlegrounds() to detect BG status transitions and
  trigger OnBattlegroundStart() when BG enters IN_PROGRESS state

- Fix crash in Map::SendObjectUpdates by removing premature bg->AddPlayer()
  call - now handled correctly by HandleMoveWorldPortAck after teleport

- Remove redundant m_bgData.queueId assignment (SetBattlegroundId already
  sets it, fixing protected member access compile error)

- Add BattlegroundCoordinatorManager for centralized BG coordination

- Update all 14 BG scripts with API compatibility fixes and improvements

+++ src/modules/Playerbot/Core/DI/Interfaces/IBattlegroundAI.h
@@ -27,7 +27,8 @@ struct Position;
 namespace Playerbot
 {
 
-enum class BGType : uint8;
+// Forward declarations - full types in BGState.h and BattlegroundAI.h
+enum class BGType : uint32;  // Map IDs (uint32)
 enum class BGRole : uint8;
 struct BGObjective;
 struct BGStrategyProfile;
